requirements.txt
asgiref==3.8.1
attrs==24.3.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
colorama==0.4.6
cryptography==44.0.0
defusedxml==0.8.0rc2
Django==5.1.1
django-bootstrap5==24.3
django-filter==24.3
django-taggit==6.1.0
djangorestframework==3.15.2
drf-spectacular==0.28.0
flake8==7.1.1
idna==3.10
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
Markdown==3.7
mccabe==0.7.0
mypy==1.11.2
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==24.1
Pillow==10.1.0
pluggy==1.5.0
psycopg==3.2.3
psycopg-binary==3.2.3
pycodestyle==2.12.1
pycparser==2.22
pyflakes==3.2.0
PyJWT==2.10.1
pytest==8.3.3
python-dotenv==1.0.1
python3-openid==3.2.0
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.22.3
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
